fix(test): cast usize to u32 in make_reaper_map fixture

+++ iem-mixer/crates/iem-server/src/backup_restore.rs
@@ -1233,7 +1233,7 @@ mod tests {
         names
             .iter()
             .enumerate()
-            .map(|(i, n)| (n.to_string(), i + 1))
+            .map(|(i, n)| (n.to_string(), (i + 1) as u32))
             .collect()
     }
